:recycle: refactor: clean up and remove generic top level files
